got a build to work but still need to fix warnings and test build failures

+++ wire5/src/commonMain/kotlin/mqtt/wire5/control/packet/format/variable/property/MaximumPacketSize.kt
@@ -5,6 +5,7 @@ package mqtt.wire5.control.packet.format.variable.property
 import kotlinx.io.core.BytePacketBuilder
 import mqtt.wire.data.Type
 
-data class MaximumPacketSize(val packetSizeLimitationBytes: UInt) : Property(0x27, Type.FOUR_BYTE_INTEGER) {
-    override fun write(bytePacketBuilder: BytePacketBuilder) = write(bytePacketBuilder, packetSizeLimitationBytes)
+data class MaximumPacketSize(val packetSizeLimitationBytes: Long) : Property(0x27, Type.FOUR_BYTE_INTEGER) {
+    override fun write(bytePacketBuilder: BytePacketBuilder) =
+        write(bytePacketBuilder, packetSizeLimitationBytes.toUInt())
 }
